Fixed jitter at low angular velocities
More in the comment below

+++ src/pc/djui/djui_hud_utils.h
@@ -24,8 +24,8 @@ enum DjuiFontType {
 };
 
 struct HudUtilsRotation {
-    s32 rotation;
-    s32 rotationDiff;
+    f32 rotation;
+    f32 rotationDiff;
     f32 prevPivotX;
     f32 prevPivotY;
     f32 pivotX;
